LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/navigation-timing-sizes.https_prefetch=true&bypass_cache=true-expected.txt


FAIL PerformanceNavigationTiming.transferSize/encodedBodySize/decodedBodySize test, same origin prefetch. assert_greater_than: expected a number greater than 0 but got 0
